Update patch set 1
Patch Set 1:

The VM it's running on is an Ubuntu image. I don't see in our packer scripts where we would be pulling in docker-py from but I wouldn't be surprised if it's getting pulled in from somewhere. I think some ODL projects use it if I recall.

Patch-set: 1
Label: Verified=0